# src/main/java/com/probestack/forgestudio/design/api/ReportsApiController.java
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.BalanceSheet;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import com.probestack.forgestudio.design.service.ReportsService;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-08T17:04:28.442050077Z[GMT]")
@Controller
@RequestMapping("${openapi.financialServices.base-path:/v1}")
public class ReportsApiController implements ReportsApi {

    private static final Logger log = LoggerFactory.getLogger(ReportsApiController.class);

    private final ReportsService reportsService;

    @Autowired()
    public ReportsApiController(ReportsService reportsService) {
        this.reportsService = reportsService;
    }

    @Override()
    public ResponseEntity<BalanceSheet> getBalanceSheet(@Valid() @RequestBody() @RequestParam() Integer fiscalYear) {
        log.info("Processing getBalanceSheet request");
        try {
            var response = reportsService.getBalanceSheet(fiscalYear);
            log.info("getBalanceSheet completed successfully");
            return response;
        } catch (Exception e) {
            log.error("Failed to process getBalanceSheet: {}", e.getMessage(), e);
            throw e;
        }
    }
}