# src/main/java/com/probestack/forgestudio/design/api/TransactionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.CreateTransactionRequest;
import org.springframework.format.annotation.DateTimeFormat;
import java.time.LocalDate;
import com.probestack.forgestudio.design.model.Transaction;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-08T17:04:28.442050077Z[GMT]")@Validated
@Tag(name = "Transactions", description = "the Transactions API")
public interface TransactionsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /transactions : Create transaction
     *
     * @param createTransactionRequest  (required)
     * @return Transaction created (status code 201)
     */
    @Operation(
        operationId = "createTransaction",
        summary = "Create transaction",
        tags = { "Transactions" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Transaction created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Transaction.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/transactions",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Transaction> createTransaction(
        @Parameter(name = "CreateTransactionRequest", description = "", required = true) @Valid @RequestBody CreateTransactionRequest createTransactionRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"transactionType\" : \"DEBIT\", \"accountId\" : \"accountId\", \"amount\" : 0.8008281904610115, \"description\" : \"description\", \"currency\" : \"currency\", \"transactionDate\" : \"2000-01-23T04:56:07.000+00:00\", \"transactionId\" : \"transactionId\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /transactions : List transactions
     *
     * @param accountId  (optional)
     * @param startDate  (optional)
     * @param endDate  (optional)
     * @return Transaction list (status code 200)
     */
    @Operation(
        operationId = "listTransactions",
        summary = "List transactions",
        tags = { "Transactions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Transaction list", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Transaction.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/transactions",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Transaction>> listTransactions(
        @Parameter(name = "accountId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "accountId", required = false) String accountId,
        @Parameter(name = "startDate", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "startDate", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate startDate,
        @Parameter(name = "endDate", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "endDate", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate endDate
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "[ { \"transactionType\" : \"DEBIT\", \"accountId\" : \"accountId\", \"amount\" : 0.8008281904610115, \"description\" : \"description\", \"currency\" : \"currency\", \"transactionDate\" : \"2000-01-23T04:56:07.000+00:00\", \"transactionId\" : \"transactionId\" }, { \"transactionType\" : \"DEBIT\", \"accountId\" : \"accountId\", \"amount\" : 0.8008281904610115, \"description\" : \"description\", \"currency\" : \"currency\", \"transactionDate\" : \"2000-01-23T04:56:07.000+00:00\", \"transactionId\" : \"transactionId\" } ]";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}