# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@example.com
    name: Finance API Team
  description: "API specification for financial account management, transactions,\
    \ payments, and reporting."
  title: Financial Services API
  version: 1.0.0
servers:
- description: Production
  url: https://api.example.com/v1
- description: Sandbox
  url: https://sandbox.example.com/v1
security:
- bearerAuth: []
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      parameters:
      - explode: true
        in: query
        name: accountId
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: startDate
        required: false
        schema:
          format: date
          type: string
        style: form
      - explode: true
        in: query
        name: endDate
        required: false
        schema:
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Transaction"
                type: array
          description: Transaction list
      summary: List transactions
      tags:
      - Transactions
      x-accepts: application/json
      x-tags:
      - tag: Transactions
    post:
      operationId: createTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
          description: Transaction created
      summary: Create transaction
      tags:
      - Transactions
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Transactions
...
  /reports/balance-sheet:
    get:
      operationId: getBalanceSheet
      parameters:
      - explode: true
        in: query
        name: fiscalYear
        required: true
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSheet"
          description: Balance sheet report
      summary: Generate balance sheet report
      tags:
      - Reports
      x-accepts: application/json
      x-tags:
      - tag: Reports
components:
  schemas:
...
    Transaction:
      example:
        transactionType: DEBIT
        accountId: accountId
        amount: 0.8008281904610115
        description: description
        currency: currency
        transactionDate: 2000-01-23T04:56:07.000+00:00
        transactionId: transactionId
      properties:
        transactionId:
          type: string
        accountId:
          type: string
        amount:
          format: double
          type: number
        currency:
          type: string
        transactionType:
          enum:
          - DEBIT
          - CREDIT
          type: string
        description:
          type: string
        transactionDate:
          format: date-time
          type: string
      type: object
    CreateTransactionRequest:
      example:
        transactionType: transactionType
        accountId: accountId
        amount: 0.8008281904610115
        description: description
      properties:
        accountId:
          type: string
        amount:
          format: double
          type: number
        transactionType:
          type: string
        description:
          type: string
      required:
      - accountId
      - amount
      - transactionType
      type: object
...
    BalanceSheet:
      example:
        assets: 6.027456183070403
        liabilities: 1.4658129805029452
        fiscalYear: 0
        equity: 5.962133916683182
      properties:
        fiscalYear:
          type: integer
        assets:
          format: double
          type: number
        liabilities:
          format: double
          type: number
        equity:
          format: double
          type: number
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http